Remove domain module with usecases. Move models and mappers to data module. Remove abstraction of repositories
