+++ app/src/main/java/io/github/msh91/arch/di/builder/ViewModelBuilder.kt
@@ -15,7 +15,6 @@ import io.github.msh91.arch.viewmodel.ArchViewModelFactory
  *
  */
 @Module(includes = [
-    (RepositoryBuilder::class),
     (HomeViewModelBuilder::class)
 ])
 abstract class ViewModelBuilder {
