+++ app/src/main/java/io/github/msh91/arch/ui/home/list/HomeListViewModel.kt
@@ -1,26 +1,48 @@
 package io.github.msh91.arch.ui.home.list
 
 import android.util.Log
-import io.github.msh91.arch.domain.model.movie.Movie
-import io.github.msh91.arch.domain.model.response.SuccessResponse
-import io.github.msh91.arch.domain.model.response.UseCaseResponse
-import io.github.msh91.arch.domain.usecase.movie.GetAllMoviesUseCase
+import io.github.msh91.arch.data.mapper.ErrorMapper
+import io.github.msh91.arch.data.model.movie.Movie
+import io.github.msh91.arch.data.model.response.ErrorResponse
+import io.github.msh91.arch.data.model.response.ErrorStatus
+import io.github.msh91.arch.data.model.response.SuccessResponse
+import io.github.msh91.arch.data.model.response.UseCaseResponse
+import io.github.msh91.arch.data.repository.movie.MovieRepository
 import io.github.msh91.arch.ui.base.BaseViewModel
 import io.github.msh91.arch.util.connectivity.BaseConnectionManager
 import io.github.msh91.arch.util.livedata.NonNullLiveData
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.schedulers.Schedulers
 import javax.inject.Inject
 
 class HomeListViewModel @Inject constructor(
         connectionManager: BaseConnectionManager,
-        private val getAllMoviesUseCase: GetAllMoviesUseCase
+        private val movieRepository: MovieRepository,
+        private val errorMapper: ErrorMapper
 ) : BaseViewModel(connectionManager) {
     private val TAG = HomeListViewModel::class.java.simpleName
 
     val movies = NonNullLiveData<List<Movie>>(emptyList())
 
     init {
-        getAllMoviesUseCase
-                .execute(compositeDisposable, this::getAllMoviesResponse, this::onTokenExpired)
+        movieRepository
+            .getAllMovies()
+            .onBackpressureLatest()
+            .subscribeOn(Schedulers.io())
+            .observeOn(AndroidSchedulers.mainThread())
+            .subscribe(
+                {
+                    getAllMoviesResponse(SuccessResponse(it))
+                },
+                {
+                    val error = errorMapper.getErrorModel(it)
+
+                    if (error.errorStatus == ErrorStatus.UNAUTHORIZED) {
+                        onTokenExpired()
+                    }
+                    getAllMoviesResponse(ErrorResponse(error))
+
+                }).also { compositeDisposable.add(it) }
     }
 
     private fun getAllMoviesResponse(response: UseCaseResponse<List<Movie>>) {
