+++ data/src/main/java/io/github/msh91/arch/data/model/response/DomainErrorException.kt
@@ -1,4 +1,4 @@
-package io.github.msh91.arch.domain.model.response
+package io.github.msh91.arch.data.model.response
 
 class DomainErrorException(val errorModel: ErrorModel): Throwable() {
 }
