+++ data/src/main/java/io/github/msh91/arch/data/model/response/ErrorModel.kt
@@ -1,4 +1,4 @@
-package io.github.msh91.arch.domain.model.response
+package io.github.msh91.arch.data.model.response
 
 /**
  * Default error model that comes from server if something goes wrong with a repository call
