+++ data/src/main/java/io/github/msh91/arch/data/source/cloud/CloudRepository.kt
@@ -2,7 +2,7 @@ package io.github.msh91.arch.data.source.cloud
 
 import io.github.msh91.arch.data.restful.APIs
 import io.github.msh91.arch.data.restful.APIsWithToken
-import io.github.msh91.arch.domain.model.movie.Movie
+import io.github.msh91.arch.data.model.movie.Movie
 import io.reactivex.Flowable
 
 /**
