+++ data/src/main/java/io/github/msh91/arch/data/source/db/AppDataBase.kt
@@ -3,7 +3,7 @@ package io.github.msh91.arch.data.source.db
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import io.github.msh91.arch.data.source.db.dao.MovieDao
-import io.github.msh91.arch.domain.model.movie.Movie
+import io.github.msh91.arch.data.model.movie.Movie
 
 @Database(entities = [Movie::class], version = AppDataBase.VERSION)
 abstract class AppDataBase : RoomDatabase() {
