--- domain/build.gradle
@@ -1,52 +1,0 @@
-import io.github.msh91.arch.Libs
-
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-kapt'
-apply plugin: 'kotlin-android-extensions'
-
-android {
-    compileSdkVersion 28
-
-
-    defaultConfig {
-        minSdkVersion 16
-        targetSdkVersion 28
-        versionCode 1
-        versionName "1.0"
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-
-    }
-    compileOptions {
-        targetCompatibility 1.8
-        sourceCompatibility 1.8
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-
-}
-
-kapt {
-    generateStubs = true
-}
-
-dependencies {
-    implementation Libs.Common.kotlin
-    implementation Libs.Jetpack.lifecycleExtensions
-    implementation Libs.Common.dagger
-    kapt Libs.Common.daggerCompiler
-    implementation Libs.Jetpack.room
-    implementation Libs.Common.gson
-    implementation Libs.Common.rxJava
-    implementation Libs.Common.rxAndroid
-    testImplementation Libs.Testing.junit
-    testImplementation Libs.Testing.mockitoKotlin
-    androidTestImplementation Libs.Testing.testRunner
-    androidTestImplementation Libs.Testing.esperesso
-}
