--- domain/src/androidTest/java/io/github/msh91/arch/domain/ExampleInstrumentedTest.java
@@ -1,27 +1,0 @@
-package io.github.msh91.arch.domain;
-
-import android.content.Context;
-
-import androidx.test.InstrumentationRegistry;
-import androidx.test.runner.AndroidJUnit4;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import static org.junit.Assert.*;
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-@RunWith(AndroidJUnit4.class)
-public class ExampleInstrumentedTest {
-  @Test
-  public void useAppContext() {
-    // Context of the app under test.
-    Context appContext = InstrumentationRegistry.getTargetContext();
-
-    assertEquals("io.github.msh91.arch.domain.test", appContext.getPackageName());
-  }
-}
