--- domain/src/main/java/io/github/msh91/arch/domain/usecase/base/LiveDataUseCase.kt
@@ -1,7 +1,0 @@
-package io.github.msh91.arch.domain.usecase.base
-
-import androidx.lifecycle.LiveData
-
-abstract class LiveDataUseCase<T> : UseCase<LiveData<T>>() {
-
-}
