--- domain/src/main/java/io/github/msh91/arch/domain/usecase/base/UseCase.kt
@@ -1,10 +1,0 @@
-package io.github.msh91.arch.domain.usecase.base
-
-/**
- * Should be used as parent class of all UseCases.
- */
-abstract class UseCase<T> {
-
-    abstract fun execute(): T
-
-}
